Allow pip package states to be controlled by vars.
If a pip package is updated, and the deployer wants to keep it
up-to-date, they would manually have to remove the package to get the
new version.

Instead, this patch allows the user to override the state so that it
might be 'latest' for the ones they want. Otherwise, the default
behavior is the same.

Fixes #5